backend/requirements.txt
# ─────────────────────────────────────────────────────────────────
# ElderMind — Python Dependencies (pinned versions)
# Install: pip install -r requirements.txt
# ─────────────────────────────────────────────────────────────────

# FastAPI core
fastapi==0.110.0
uvicorn==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.28
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Rate limiting
slowapi==0.1.9

# AI & ML
openai==1.12.0
chromadb==0.4.22          # PINNED — do NOT upgrade; 0.5+ breaks Suchit's filter syntax
sentence-transformers==2.5.1
scikit-learn==1.4.0
numpy==1.26.4

# Voice pipeline
httpx==0.27.0
pydub==0.25.1
websockets==12.0

# Scheduling & SMS
apscheduler==3.10.4
twilio==8.10.0

# Utilities
python-dotenv==1.0.1
pydantic==2.6.1
